agent-service/requirements.txt
fastapi==0.116.1
sqlmodel==0.0.24
gunicorn==23.0.0
redis==6.4.0
# For running Postgres Database
psycopg2-binary==2.9.10

python-dotenv==1.1.1
python-json-logger==3.3.0

pytest==8.4.1

# Langchain
langchain==0.3.27
# langchain-ollama==0.3.3
langchain-community==0.3.27
langchain-milvus==0.2.1
langchain-google-genai==2.1.9

langgraph==0.6.6
langgraph-cli[inmem]==0.3.8

unstructured==0.18.13
networkx==3.5
openpyxl==3.1.5

beautifulsoup4==4.13.4

pandas==2.3.2

nltk==3.9.1

docling==2.47.0

minio==7.2.16
